+++ .storybook/main.js
@@ -19,4 +19,5 @@ module.exports = {
 		autodocs: 'tag',
 	},
 	typescript: {reactDocgen: false},
+	staticDirs: ['../src/fonts/'],
 };